Registered the Videos model to the videos app admin
